requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-csp==3.8
whitenoise==6.6.0
gunicorn==21.2.0
Pillow==10.3.0
openpyxl==3.1.2
python-slugify==8.0.4

pandas==2.2.2
